refactor(notes): extract buildMarkdown to lib/notesMarkdown to fix Next.js page export error
Next.js rejects arbitrary named exports from page files. Moved buildMarkdown,
chapterLabel, and truncate to src/lib/notesMarkdown.ts; notes page imports
only what it needs. Coverage test now imports buildMarkdown from the library.
